+++ src/main/java/ru/hh/spb/computershop/data/ProductType.java
@@ -1,0 +1,19 @@
+package ru.hh.spb.computershop.data;
+
+public enum ProductType implements java.io.Serializable{
+
+    UNKNOWN,
+    COMPUTER,
+    NOTEBOOK,
+    MONITOR,
+    HDD;
+
+    public static ProductType fromValue(String value) {
+        for (ProductType type : values()) {
+            if (type.name().equalsIgnoreCase(value)) {
+                return type;
+            }
+        }
+        return UNKNOWN;
+    }
+}
