+++ src/main/java/ru/hh/spb/computershop/entities/Computer.java
@@ -2,9 +2,9 @@ package ru.hh.spb.computershop.entities;
 
 import jakarta.persistence.Entity;
 
-import ru.hh.spb.computershop.enums.ComputerType;
-import ru.hh.spb.computershop.enums.Manufacturer;
-import ru.hh.spb.computershop.enums.ProductType;
+import ru.hh.spb.computershop.data.ComputerType;
+import ru.hh.spb.computershop.data.Manufacturer;
+import ru.hh.spb.computershop.data.ProductType;
 
 @Entity
 public class Computer extends Product {
