+++ src/main/java/ru/hh/spb/computershop/entities/Monitor.java
@@ -2,7 +2,7 @@ package ru.hh.spb.computershop.entities;
 
 import jakarta.persistence.Entity;
 
-import ru.hh.spb.computershop.enums.Manufacturer;
+import ru.hh.spb.computershop.data.Manufacturer;
 
 @Entity
 public class Monitor extends Product {
