--- src/main/java/ru/hh/spb/computershop/enums/ComputerType.java
@@ -1,9 +1,0 @@
-package ru.hh.spb.computershop.enums;
-
-public enum ComputerType implements java.io.Serializable {
-
-    UNKNOWN,
-    DESKTOP,
-    NETTOP,
-    MONOBLOCK,
-}
