--- src/main/java/ru/hh/spb/computershop/enums/Manufacturer.java
@@ -1,17 +1,0 @@
-package ru.hh.spb.computershop.enums;
-
-public enum Manufacturer implements java.io.Serializable {
-
-    UNKNOWN,
-    ASUS,
-    ACER,
-    APPLE,
-    SAMSUNG,
-    WESTERN_DIGITAL,
-    SEAGATE,
-    TOSHIBA,
-    AOC,
-    LG,
-    SONY,
-    MICROSOFT,
-}
