--- src/main/java/ru/hh/spb/computershop/enums/ProductType.java
@@ -1,10 +1,0 @@
-package ru.hh.spb.computershop.enums;
-
-public enum ProductType implements java.io.Serializable{
-
-    UNKNOWN,
-    COMPUTER,
-    NOTEBOOK,
-    MONITOR,
-    HDD,
-}
