+++ src/main/java/ru/hh/spb/computershop/responses/SuccessfulResponse.java
@@ -1,6 +1,6 @@
 package ru.hh.spb.computershop.responses;
 
-import ru.hh.spb.computershop.enums.ResponseType;
+import ru.hh.spb.computershop.data.ResponseType;
 
 public record SuccessfulResponse(ResponseType type, Object data, String timestamp) implements ShopResponse {
 }
