+++ src/main/java/ru/hh/spb/computershop/services/ProductService.java
@@ -8,10 +8,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import ru.hh.spb.computershop.entities.*;
-import ru.hh.spb.computershop.enums.ComputerType;
-import ru.hh.spb.computershop.enums.Manufacturer;
-import ru.hh.spb.computershop.enums.NotebookSize;
-import ru.hh.spb.computershop.enums.ProductType;
+import ru.hh.spb.computershop.data.ComputerType;
+import ru.hh.spb.computershop.data.Manufacturer;
+import ru.hh.spb.computershop.data.NotebookSize;
+import ru.hh.spb.computershop.data.ProductType;
 import ru.hh.spb.computershop.repositories.ProductRepository;
 
 import java.util.List;
